chore: remove Application Insights and Azure App Configuration
