+++ .ci/appsettings.override.postgres.json
@@ -109,7 +109,6 @@
         }
     },
     "Serilog": {
-        "Using": ["Serilog.Sinks.ApplicationInsights"],
         "MinimumLevel": {
             "Default": "Debug"
         },
