--- ConsumerApi/ApplicationInsights/TelemetryInitializers/CloudRoleNameTelemetryInitializer.cs
@@ -1,13 +1,0 @@
-﻿using Microsoft.ApplicationInsights.Channel;
-using Microsoft.ApplicationInsights.Extensibility;
-
-namespace ConsumerApi.ApplicationInsights.TelemetryInitializers;
-
-public class CloudRoleNameTelemetryInitializer : ITelemetryInitializer
-{
-    public void Initialize(ITelemetry telemetry)
-    {
-        if (string.IsNullOrEmpty(telemetry.Context.Cloud.RoleName))
-            telemetry.Context.Cloud.RoleName = "Challenges";
-    }
-}
