+++ ConsumerApi/ConsumerApi.csproj
@@ -14,17 +14,14 @@
 		<PackageReference Include="AspNetCore.HealthChecks.NpgSql" Version="[6.0.2]" />
 		<PackageReference Include="AspNetCore.HealthChecks.SqlServer" Version="7.0.0" />
 		<PackageReference Include="FluentValidation.AspNetCore" Version="11.3.0" />
-		<PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.21.0" />
 		<PackageReference Include="Microsoft.AspNet.WebApi.Client" Version="5.2.9" />
 		<PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="7.0.12" />
 		<PackageReference Include="NetEscapades.AspNetCore.SecurityHeaders" Version="0.21.0" />
 		<PackageReference Include="Autofac.Extensions.DependencyInjection" Version="8.0.0" />
-		<PackageReference Include="Azure.Identity" Version="1.10.2" />
 		<PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="7.0.12" />
 		<PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.19.5" />
 		<PackageReference Include="ReHackt.Extensions.Options.Validation" Version="7.0.1" />
 		<PackageReference Include="Swashbuckle.AspNetCore" Version="6.5.0" />
-		<PackageReference Include="Microsoft.Extensions.Configuration.AzureAppConfiguration" Version="6.1.0" />
 		<PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="7.0.12" />
 		<PackageReference Include="System.Data.SqlClient" Version="4.8.5" />
 		<PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="7.0.12">
@@ -41,7 +38,7 @@
 		<PackageReference Include="Serilog.Exceptions.EntityFrameworkCore" Version="8.4.0" />
 		<PackageReference Include="Serilog.Formatting.Compact" Version="1.1.0" />
 		<PackageReference Include="Serilog.Sinks.Http" Version="8.0.0" />
-        <PackageReference Include="Serilog.Sinks.Seq" Version="5.2.3" />
+		<PackageReference Include="Serilog.Sinks.Seq" Version="5.2.3" />
 	</ItemGroup>
 
 	<ItemGroup>
