+++ Modules/Devices/src/Devices.Infrastructure/PushNotifications/AzureNotificationHub/NotificationBuilder.cs
@@ -1,6 +1,4 @@
 ﻿using System.Text.Json;
-using System.Text.Json.Serialization;
-using Enmeshed.Tooling.Extensions;
 using Microsoft.Azure.NotificationHubs;
 
 namespace Backbone.Modules.Devices.Infrastructure.PushNotifications.AzureNotificationHub;
@@ -35,17 +33,4 @@ public abstract class NotificationBuilder
     }
 
     public abstract Notification Build();
-}
-
-public class DateTimeConverter : JsonConverter<DateTime>
-{
-    public override DateTime Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-    {
-        throw new NotImplementedException();
-    }
-
-    public override void Write(Utf8JsonWriter writer, DateTime value, JsonSerializerOptions options)
-    {
-        writer.WriteStringValue(value.ToUniversalString());
-    }
 }
